--- CapstoneProject/src/main/java/com/example/rechargeapp/AdminController.java
@@ -1,5 +1,0 @@
-package com.example.rechargeapp;
-
-public class AdminController {
-
-}
