--- CapstoneProject/src/main/java/com/example/rechargeapp/PaymentController.java
@@ -1,5 +1,0 @@
-package com.example.rechargeapp;
-
-public class PaymentController {
-
-}
