--- CapstoneProject/src/main/java/com/example/rechargeapp/PlanController.java
@@ -1,5 +1,0 @@
-package com.example.rechargeapp;
-
-public class PlanController {
-
-}
