+++ CapstoneProject/src/main/java/com/example/rechargeapp/controller/AdminController.java
@@ -1,0 +1,5 @@
+package com.example.rechargeapp.controller;
+
+public class AdminController {
+
+}
