+++ CapstoneProject/src/main/java/com/example/rechargeapp/controller/CustomerController.java
@@ -1,4 +1,4 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.controller;
 
 import javax.validation.Valid;
 
@@ -7,6 +7,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.example.rechargeapp.Status;
+import com.example.rechargeapp.models.Customer;
+import com.example.rechargeapp.service.CustomerService;
 
 import antlr.collections.List;
