+++ CapstoneProject/src/main/java/com/example/rechargeapp/controller/PlanController.java
@@ -1,0 +1,5 @@
+package com.example.rechargeapp.controller;
+
+public class PlanController {
+
+}
