+++ CapstoneProject/src/main/java/com/example/rechargeapp/models/Customer.java
@@ -1,5 +1,5 @@
 
-package com.example.rechargeapp;
+package com.example.rechargeapp.models;
 
 import java.util.Date;
