+++ CapstoneProject/src/main/java/com/example/rechargeapp/models/Offers.java
@@ -1,4 +1,4 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.models;
 
 import java.util.Date;
