+++ CapstoneProject/src/main/java/com/example/rechargeapp/repository/AdminRepository.java
@@ -1,6 +1,8 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.rechargeapp.models.Admin;
 
 public interface AdminRepository extends JpaRepository<Admin, Integer>{
