+++ CapstoneProject/src/main/java/com/example/rechargeapp/repository/CustomerRepository.java
@@ -1,7 +1,9 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.example.rechargeapp.models.Customer;
 
 @Repository
 public interface CustomerRepository extends JpaRepository<Customer, Integer> {
