+++ CapstoneProject/src/main/java/com/example/rechargeapp/repository/PlanRepository.java
@@ -1,10 +1,12 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.example.rechargeapp.models.Plan;
+
 public interface PlanRepository extends JpaRepository<Plan, Integer>{
 
-	Plan findByPaymentId(int planId);
+	Plan findByPlanId(int planId);
 
 }
