+++ CapstoneProject/src/main/java/com/example/rechargeapp/service/AdminService.java
@@ -1,7 +1,10 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.example.rechargeapp.models.Admin;
+import com.example.rechargeapp.repository.AdminRepository;
 
 @Service
 public class AdminService {
