+++ CapstoneProject/src/main/java/com/example/rechargeapp/service/PlanService.java
@@ -1,7 +1,11 @@
-package com.example.rechargeapp;
+package com.example.rechargeapp.service;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.example.rechargeapp.models.Plan;
+import com.example.rechargeapp.repository.PlanRepository;
 @Service
 public class PlanService {
 	@Autowired
