+++ CapstoneProject/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 
 spring.datasource.url=jdbc:mysql://localhost:3306/rechargeappdb?useSSL=false
 spring.datasource.username=root
-spring.datasource.password=123456789
+spring.datasource.password=root123
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql=true
